+++ src/main/kotlin/com/petqua/common/exception/auth/OauthClientExceptionType.kt
@@ -1,4 +1,4 @@
-package com.petqua.common.exception.oauth
+package com.petqua.common.exception.auth
 
 import com.petqua.common.exception.BaseExceptionType
 import org.springframework.http.HttpStatus
