+++ src/main/kotlin/com/petqua/domain/auth/AuthToken.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.oauth
+package com.petqua.domain.auth
 
 class AuthToken(
     val accessToken: String,
