+++ src/main/kotlin/com/petqua/domain/auth/OauthClient.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.oauth
+package com.petqua.domain.auth
 
 interface OauthClient {
