+++ src/main/kotlin/com/petqua/domain/auth/OauthTokenInfo.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.oauth
+package com.petqua.domain.auth
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming
