+++ src/main/kotlin/com/petqua/presentation/OauthServerTypeConverter.kt
@@ -1,6 +1,6 @@
 package com.petqua.presentation
 
-import com.petqua.domain.oauth.OauthServerType
+import com.petqua.domain.auth.OauthServerType
 import org.springframework.core.convert.converter.Converter
 
 class OauthServerTypeConverter : Converter<String, OauthServerType> {
